chore: refactor and colocate types
