+++ src/features/products/Products.tsx
@@ -1,8 +1,12 @@
 'use client';
 import { useQuery } from '@tanstack/react-query';
 import { getProducts } from '@/store/products.get';
-import type { Product } from '@/types';
 import ProductsTable from './ui/ProductsTable';
+import type { Product as ProductType, Unit } from '@prisma/client';
+
+export type Product = ProductType & {
+  units: Unit[];
+};
 
 export default function Products() {
   const { isLoading, isError, data, error } = useQuery<Product[], Error>({
