+++ src/store/products.get.ts
@@ -1,5 +1,4 @@
-// TODO: fix /types
-import { Product } from '@/types';
+import { Product } from '@/features/products/Products';
 
 export const getProducts = async (): Promise<Product[]> => {
   const dataList: Product[] = await fetch('/api/get-products').then((res) =>
